(fix) kucoin unit tests passing
