+++ hummingbot/connector/constants.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 NaN = float("nan")
 s_decimal_NaN = Decimal("nan")
 s_decimal_0 = Decimal(0)
